requirements.txt
python-telegram-bot==13.15
requests
names
Faker
fake_useragent